+++ worker/src/main/java/org/apache/celeborn/service/deploy/worker/storage/CreditStreamManager.java
@@ -78,9 +78,16 @@ public class CreditStreamManager {
       int initialCredit,
       int startSubIndex,
       int endSubIndex,
-      FileInfo fileInfo) throws IOException {
+      FileInfo fileInfo)
+      throws IOException {
     return registerStream(
-        notifyStreamHandlerCallback, channel, initialCredit, startSubIndex, endSubIndex, fileInfo, false);
+        notifyStreamHandlerCallback,
+        channel,
+        initialCredit,
+        startSubIndex,
+        endSubIndex,
+        fileInfo,
+        false);
   }
 
   public long registerStream(
